requirements.txt
flask==3.0.2
google-genai==0.3.0
python-dotenv==1.0.1
